Add RSpec matcher to compare arrays with a delta
Since I keep getting bogus test failures due to small floating-point
errors. This should give me a consistent way to prevent those.

Supposed to be diffable but somehow it doesn't work -- I think this
might be something to do with the delta as an extra argument.
Maybe I need a fluent interface or something?