Draft: Account for timing in polyrhythm beats
This works for timing, but I'm now not properly calculating
durations for silent beats.

That should be a fairly easy fix -- including zero amplitudes
in the canonical form -- but I don't want to go through with
it without more thought about what this implies.

Normally we stop a beat early in a polyrhythm because the next
beat's interrupted it; but if that next beat is silent, shouldn't
we continue the full length?