Set the tonic only when generating audio
Or not at all if generating MIDI, since the tonic is part of the tuning
that might feasibly vary at time of performance.

This also fixes a bug with notes in generated MIDI files being way out
of tune. I really need better tests that would have picked this up
(sending a different format in tests and real code).

We don't yet allow setting the tonic as an option, but it's feasible
now.