Switch to using JSON library for export
And begin exporting musical form vocals in a better format.
As I do this I'm realising the data format isn't fully mapped yet,
so I may end up needing to contribute to the field analysis to get
this to work properly — but hopefully the basic data Cacophony already
uses will be there.