Export scale data as well as musical forms
This is a start, but unfortunately I don't think it'll be portable.
Several of the fields I'm using are unnamed in the version of dfhack
I've got, and apparently the `anon_n` field names aren't stable.

I need either a newer DFHack, a PR to them that labels the fields
properly, or both.