Use exact MIDI note values where possible
When we can, we want to use the exact MIDI note values for our
frequencies, so they play correctly without retuning.

If the scale has more than 12 notes, or if they're all bunched at one
end, we're out of luck. But if not, we can at least get some notes to
use the correct 12TET frequency by default — and in the common-sh case
where DF bases the octave on a 12-note semitone scale, this actually is
12TET and will work fine.

For scales that aren't exactly 12TET, like a 7-note one, we only sort of
try to match notes up. We could try a bit harder to get close
frequencies, but because of beating/"out of tune" sounds, "almost the
right frequency" isn't really very useful — and 2^1/7 and 2^1/12 aren't
in an integer ratio, so perfect matches aren't possible in those cases.

This does create weird doubled or "backwards" tuning files in some
cases. That's not a problem for music we intend them for, which doesn't
use those warped tunings. If it's an annoyance in future we can always
add code to fill these in better — or an option to turn off matching to
12TET altogether.