Calculate chords based on perfect-fourth divisions
Assuming (a) that the perfect fourth divisions are all equal-sized, and
(b) that, unlike octave divisions, a division of the perfect fourth into
N notes has the Nth note at a 4:3 ratio (whereas the Nth note of an
octave division is 2^(1/12) below the octave).

I can't see any grounds to assume (b), except that without it it's a lot
less clear how notes fit together into a scale. In particular, it seems
like the scales without this assumption wouldn't actually include the
perfect fourth above the octave. Since it's such a good ratio, this
seems odd, especially for a culture that considers the perfect fourth
division fundamental.