Build scales from perfect-fourth-based chords
By transposing one of the chords up.
I'm not sure we need to do this --- it seems like DF creates "low"
perfect-fourth chords that only show up as the first half of a scale,
and "high" chords only used as the second half, but doesn't mix them.

But I'm not 100% sure (absence of evidence, not evidence of absence),
and if it ever does use one chord in both positions we'd need this
transposing feature, so I'm going with the transposition behaviour
rather than trying to guess which chords are meant to be high or low.